Clean up and move some methods to util
