+++ app/src/main/java/io/homeassistant/companion/android/vehicle/MainVehicleScreen.kt
@@ -9,9 +9,6 @@ import androidx.annotation.RequiresApi
 import androidx.car.app.CarContext
 import androidx.car.app.Screen
 import androidx.car.app.model.Action
-import androidx.car.app.model.CarColor
-import androidx.car.app.model.CarIcon
-import androidx.car.app.model.GridItem
 import androidx.car.app.model.GridTemplate
 import androidx.car.app.model.ItemList
 import androidx.car.app.model.Template
@@ -20,18 +17,15 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import com.mikepenz.iconics.IconicsDrawable
-import com.mikepenz.iconics.typeface.library.community.material.CommunityMaterial
-import com.mikepenz.iconics.utils.sizeDp
-import com.mikepenz.iconics.utils.toAndroidIconCompat
 import io.homeassistant.companion.android.BuildConfig
 import io.homeassistant.companion.android.common.data.authentication.SessionState
 import io.homeassistant.companion.android.common.data.integration.Entity
 import io.homeassistant.companion.android.common.data.integration.domain
-import io.homeassistant.companion.android.common.data.integration.getIcon
 import io.homeassistant.companion.android.common.data.prefs.PrefsRepository
 import io.homeassistant.companion.android.common.data.servers.ServerManager
-import io.homeassistant.companion.android.common.util.capitalize
+import io.homeassistant.companion.android.util.vehicle.getChangeServerGridItem
+import io.homeassistant.companion.android.util.vehicle.getDomainList
+import io.homeassistant.companion.android.util.vehicle.getNavigationGridItem
 import io.homeassistant.companion.android.util.vehicle.nativeModeActionStrip
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
@@ -39,8 +33,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
-import java.util.Calendar
-import java.util.Locale
 import io.homeassistant.companion.android.common.R as commonR
 
 @RequiresApi(Build.VERSION_CODES.O)
@@ -56,7 +48,7 @@ class MainVehicleScreen(
     companion object {
         private const val TAG = "MainVehicleScreen"
 
-        private val SUPPORTED_DOMAINS_WITH_STRING = mapOf(
+        val SUPPORTED_DOMAINS_WITH_STRING = mapOf(
             "button" to commonR.string.buttons,
             "cover" to commonR.string.covers,
             "input_boolean" to commonR.string.input_booleans,
@@ -164,68 +156,34 @@ class MainVehicleScreen(
         } else {
             var builder = ItemList.Builder()
             if (domains.isNotEmpty()) {
-                builder = addDomainList(domains)
+                builder = getDomainList(
+                    domains,
+                    carContext,
+                    screenManager,
+                    serverManager,
+                    serverId,
+                    prefsRepository,
+                    allEntities
+                )
             }
 
             builder.addItem(
-                GridItem.Builder()
-                    .setImage(
-                        CarIcon.Builder(
-                            IconicsDrawable(
-                                carContext,
-                                CommunityMaterial.Icon3.cmd_map_outline
-                            ).apply {
-                                sizeDp = 48
-                            }.toAndroidIconCompat()
-                        )
-                            .setTint(CarColor.DEFAULT)
-                            .build()
-                    )
-                    .setTitle(carContext.getString(commonR.string.aa_navigation))
-                    .setOnClickListener {
-                        Log.i(TAG, "Navigation clicked")
-                        screenManager.push(
-                            MapVehicleScreen(
-                                carContext,
-                                serverManager.integrationRepository(serverId.value),
-                                allEntities.map { it.values.filter { entity -> entity.domain in MAP_DOMAINS } }
-                            )
-                        )
-                    }
-                    .build()
+                getNavigationGridItem(
+                    carContext,
+                    screenManager,
+                    serverManager.integrationRepository(serverId.value),
+                    allEntities
+                ).build()
             )
 
             if (serverManager.defaultServers.size > 1) {
                 builder.addItem(
-                    GridItem.Builder()
-                        .setImage(
-                            CarIcon.Builder(
-                                IconicsDrawable(
-                                    carContext,
-                                    CommunityMaterial.Icon2.cmd_home_switch
-                                ).apply {
-                                    sizeDp = 48
-                                }.toAndroidIconCompat()
-                            )
-                                .setTint(CarColor.DEFAULT)
-                                .build()
-                        )
-                        .setTitle(carContext.getString(commonR.string.aa_change_server))
-                        .setOnClickListener {
-                            Log.i(TAG, "Change server clicked")
-                            screenManager.pushForResult(
-                                ChangeServerScreen(
-                                    carContext,
-                                    serverManager,
-                                    serverId
-                                )
-                            ) {
-                                it?.toString()?.toIntOrNull()?.let { serverId ->
-                                    onChangeServer(serverId)
-                                }
-                            }
-                        }
-                        .build()
+                    getChangeServerGridItem(
+                        carContext,
+                        screenManager,
+                        serverManager,
+                        serverId
+                    ) { onChangeServer(it) }.build()
                 )
             }
             builder
@@ -258,60 +216,5 @@ class MainVehicleScreen(
                 }
             carRestrictionManager?.registerListener(listener)
         }
-    }
-
-    fun addDomainList(domains: MutableSet<String>): ItemList.Builder {
-        val listBuilder = ItemList.Builder()
-        domains.forEach { domain ->
-            val friendlyDomain =
-                SUPPORTED_DOMAINS_WITH_STRING[domain]?.let { carContext.getString(it) }
-                    ?: domain.split("_").joinToString(" ") { word ->
-                        word.capitalize(Locale.getDefault())
-                    }
-            val icon = Entity(
-                "$domain.ha_android_placeholder",
-                "",
-                mapOf<Any, Any>(),
-                Calendar.getInstance(),
-                Calendar.getInstance(),
-                null
-            ).getIcon(carContext)
-
-            listBuilder.addItem(
-                GridItem.Builder().apply {
-                    if (icon != null) {
-                        setImage(
-                            CarIcon.Builder(
-                                IconicsDrawable(carContext, icon)
-                                    .apply {
-                                        sizeDp = 48
-                                    }.toAndroidIconCompat()
-                            )
-                                .setTint(CarColor.DEFAULT)
-                                .build()
-                        )
-                    }
-                }
-                    .setTitle(friendlyDomain)
-                    .setOnClickListener {
-                        Log.i(TAG, "Domain:$domain clicked")
-                        screenManager.push(
-                            EntityGridVehicleScreen(
-                                carContext,
-                                serverManager,
-                                serverId,
-                                prefsRepository,
-                                serverManager.integrationRepository(serverId.value),
-                                friendlyDomain,
-                                domains,
-                                allEntities.map { it.values.filter { entity -> entity.domain == domain } },
-                                allEntities
-                            ) { }
-                        )
-                    }
-                    .build()
-            )
-        }
-        return listBuilder
     }
 }
